\name{ASPBay-package}
\alias{ASPBay-package}
\alias{ASPBay}
\docType{package}
\title{ Bayesian method with affected sib pairs }
\description{
This package permits to select in a genomic region a subset of SNPs which is likely to contain the true causal SNPs or a SNPs which tag them.
Then, we exploite the linkage information contained in affected sib-pairs data to make inference on the causal SNPs in the region using Bayesian method.}
\author{ Claire Dandine-Roulland}
\references{ Perdry, Herve, Muller-Myhsok, Bertram, et Clerget-Darpoux, Francoise. \emph{Using affected sib-pairs to uncover rare disease variants.} Hum Hered, 2013.

Dandine-Roulland, Claire and Perdry, Herve. \emph{Where is the causal variant? On the advantage of the family design over the case-control design in genetic association studies.} 
Submitted to Eur J Hum Genet}
\keyword{ Bayesian }
\keyword{ Affected sib pairs }
